% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr.R
\docType{package}
\name{floR-package}
\alias{floR-package}
\title{floR package}
\description{
floR is an R package to implement circuit theory-based connectivity for dynamic meta-population models
}
\details{
#' @useDynLib floR
#' @importFrom Rcpp sourceCpp
NULL
}

