% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr.R
\name{flowr}
\alias{flowr}
\title{flowr objects}
\usage{
flowr(habitat, populations, output_format = "asc", output_dir = "./",
  cores = 4, control = list())
}
\arguments{
\item{habitat}{A habitat map or resistance surface, this need to be a raster (.asc).}

\item{populations}{A set focal nodes or source/destination populations (.txt list of point pairs to calculate or .asc grid).
Inputs must be points.}

\item{output_format}{The required output format, the default is a raster.}

\item{output_dir}{Directory to save output results.}

\item{cores}{The numer of processes to run in parallel. Default is four.}

\item{control}{A list of optional calls, see details for instructions.}
}
\description{
A wrapper function around the groovy 'GFlow' software to using in dynamic metapopulation models.
}
\details{
Control parameters that can be called as optional extras.
\itemize{
 \item{output_sum_density_filename}{Logical Set Output Path, file name, and format of cumulative current density. Omitting this flag will discard current density summation.}
 \item{node_pairs}{Calculate only desired node pairs if input (e.g., "${OUTPUT_DIR}/shuf.tsv" from gshuf above). Currently not used.}
 \item{converge_at}{Set Convergence Factor to stop calculating. Typically used in place of 'node_pairs' or if all pairwise is too computationlly time consuing. Acceptable formats include: '4N' or '.9999'. Set to '1N' Below.}
 \item{shuffle_node_pairs}{Shuffles pairs for random selection. Input is binary. Currently set to shuffle below (= 1)}
 \item{effective_resistance}{Print effective resistance to log file. Supply path for .csv}
}
}

